+++ app/src/main/java/org/simple/clinic/questionnaire/component/InputFieldComponentData.kt
@@ -5,6 +5,7 @@ import com.squareup.moshi.JsonClass
 import kotlinx.parcelize.Parcelize
 import org.simple.clinic.questionnaire.component.properties.InputFieldType
 import org.simple.clinic.questionnaire.component.properties.InputFieldValidations
+import org.simple.clinic.questionnaire.component.properties.InputFieldViewType
 
 @JsonClass(generateAdapter = true)
 @Parcelize
@@ -20,6 +21,12 @@ data class InputFieldComponentData(
 
     @Json(name = "type")
     val type: InputFieldType,
+
+    @Json(name = "view_type")
+    val viewType: InputFieldViewType?,
+
+    @Json(name = "view_format")
+    val viewFormat: String?,
 
     @Json(name = "validations")
     val validations: InputFieldValidations