+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Show progress indicator when logging out
 - Bump to latest alpha/beta version of Compose libraries
 - Add month and year picker dialog for Compose forms
+- Add support for date input field for Compose forms
 
 ### Changes